Side Collision Update
Side collision is implemented and nearly fully functional. It isn't
perfect, but the solution does not hinder game controls or break the
game in any major way. It's after midnight at time of commit. I will
rise early tomorrow and add a quick commit with more comments describing
the changes, along with fixes to the collision system.